requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.8.1
bcrypt==4.2.1
binaryornot==0.4.4
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
cookiecutter==2.6.0
crispy-bootstrap5==2025.4
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.1
django-appconf==1.1.0
django-bootstrap5==24.2
django-crispy-forms==2.3
django-environ==0.12.0
django-import-export==4.2.0
django-select2==8.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
Faker==36.1.1
fastapi==0.115.8
flet==0.26.0
flet-cli==0.26.0
flet-desktop==0.26.0
flet-web==0.26.0
fpdf==1.7.2
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
humanize==4.10.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.1.2
oauthlib==3.2.2
odfpy==1.4.1
openai==1.61.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pefile==2024.8.26
pillow==11.0.0
postgres==4.0
pptxtopdf==0.0.2
proto-plus==1.26.1
protobuf==6.30.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
psycopg2-pool==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.2.3
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzipper==0.3.6
qrcode==7.4.2
repath==0.9.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.45.3
tablib==3.7.0
tabulate==0.9.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==4.0.2
watchfiles==1.0.4
websockets==15.0
whitenoise==6.9.0
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
